rework build script

+++ build.sh
@@ -14,19 +14,14 @@ CMAKE_GENERATOR_LOCAL=-G"Eclipse CDT4 - Unix Makefiles"
 
 usage()
 {
-   echo "build.sh <Command>"
+   echo "build.sh (info | local | deb | rpm | tst | clean)"
    echo ""
-   echo "Commands:"
-   echo "  info" 
-   echo "    create version and dependency files"
-   echo "  local [renew]"
-   echo "    create or renew local build and install in folder ${HOME}/install"
-   echo "  deb [renew]"
-   echo "    create or renew deb package"
-   echo "  rpm"
-   echo "    create or renew rpm package"
-   echo "  clean"
-   echo "    delete all build directories"
+   echo "  info  - create version and dependency files"
+   echo "  local - create local build and install in folder ${HOME}/install"
+   echo "  deb   - create deb package"
+   echo "  rpm   - create rpm package"
+   echo "  tst   - build unit application"
+   echo "  clean - delete all build directories"
 }
 
 
@@ -37,7 +32,7 @@ usage()
 #
 # -----------------------------------------------------------------------------
 # -----------------------------------------------------------------------------
-version_info()
+build_info()
 {
     echo "version info"
 
@@ -47,7 +42,7 @@ version_info()
     cmake -DCREATE_INFO_FILES=1 ../src/
 }
 
-version_info_clean()
+build_info_clean()
 {
     rm -rf build_info
 }
@@ -64,54 +59,39 @@ build_local()
 {
     echo "build local start"
 
-    # build local directory
-    rm -rf build_local
-    mkdir build_local
-    cd build_local
-
-    # build local
-    set -x
-    cmake ../src \
-          "${CMAKE_GENERATOR_LOCAL}" 
-    set +x
-    if [ $? -ne 0 ] ;
+    # check build directoriy
+    if [ ! -d "build_local" ] ;
     then
-        echo "cmake error"
-        return $?
-    fi
-
-    # install local
-    make DESTDIR="${HOME}/install" install
-     if [ $? -ne 0 ] ;
-    then
-        echo "make install error"
-        return $?
-    fi
-
-    return 0
-}
-
-
-build_local_renew()
-{
-    echo "renew local start"
-
-    # build local directory
-    if [ ! -d "build_local" ] ; 
-    then
-        echo "error - directory build_local do not exist"
-        return 1;
+        BUILD_FIRST=1
+        rm -rf build_local
+        mkdir build_local
+    else
+        BUILD_FIRST=0
     fi
     cd build_local
 
     # build local
-    set -x
-    cmake .
-    set +x
-    if [ $? -ne 0 ] ;
+    if [ ${BUILD_FIRST} -eq 1 ] ;
     then
-        echo "cmake error"
-        return $?
+        set -x
+        cmake ../src \
+              "${CMAKE_GENERATOR_LOCAL}" 
+        set +x
+        if [ $? -ne 0 ] ;
+        then
+            echo "cmake error"
+            return $?
+        fi
+    else
+        set -x
+        cmake .
+        set +x
+        if [ $? -ne 0 ] ;
+        then
+            echo "cmake error"
+            return $?
+        fi
+
     fi
 
     # install local
@@ -152,21 +132,34 @@ build_deb()
         return -1
     fi
 
-    # build package directory
-    rm -rf build_deb
-    mkdir build_deb
+    # check build directoriy
+    if [ ! -d "build_deb" ] ;
+    then
+        BUILD_FIRST=1
+        rm -rf build_deb
+        mkdir build_deb
+    else
+        BUILD_FIRST=0
+    fi
     cd build_deb
 
     # build package
-    cmake ../src \
-        "${CMAKE_GENERATOR_LOCAL}" \
-        "-DCPACK_BINARY_DEB=1" \
-        "-DCPACK_BINARY_RPM=0" 
-    if [ $? -ne 0 ] ;
+    if [ ${BUILD_FIRST} -eq 1 ] ;
     then
-        echo "cmake error"
-        return $?
+ 
+        cmake ../src \
+            "${CMAKE_GENERATOR_LOCAL}" \
+            "-DCPACK_BINARY_DEB=1" \
+            "-DCPACK_BINARY_RPM=0" 
+        if [ $? -ne 0 ] ;
+        then
+            echo "cmake error"
+            return $?
+        fi
+    else
+        cmake .
     fi
+
     make package
     if [ $? -ne 0 ] ;
     then
@@ -176,36 +169,6 @@ build_deb()
 
     return 0
 }
-
-build_deb_renew()
-{
-    echo "renew deb package start"
-
-    # build package directory
-    if [ ! -d "build_deb" ] ; 
-    then
-        echo "error - directory build_deb do not exist"
-        return 1;
-    fi
-    cd build_deb
-
-    # build package
-    cmake .
-    if [ $? -ne 0 ] ;
-    then
-        echo "cmake error"
-        return $?
-    fi
-    make package
-    if [ $? -ne 0 ] ;
-    then
-        echo "make package error"
-        return $?
-    fi
-
-    return 0
-}
-
 
 build_deb_clean()
 {
@@ -234,21 +197,32 @@ build_rpm()
     fi
     
     # build package directory
-    rm -rf build_rpm
-    mkdir build_rpm
+    if [ ! -d "build_rpm" ] ;
+    then
+        BUILD_FIRST=1
+        rm -rf build_rpm
+        mkdir build_rpm
+    else
+        BUILD_FIRST=0
+    fi
     cd build_rpm
 
     # build package
-   echo "${CPACK_BINARY}"
-    cmake ../src \
-        "${CMAKE_GENERATOR_LOCAL}" \
-        "-DCPACK_BINARY_DEB=0" \
-        "-DCPACK_BINARY_RPM=1"
-    if [ $? -ne 0 ] ;
+    if [ ${BUILD_FIRST} -eq 1 ] ;
     then
-        echo "cmake error"
-        return $?
+        cmake ../src \
+            "${CMAKE_GENERATOR_LOCAL}" \
+            "-DCPACK_BINARY_DEB=0" \
+            "-DCPACK_BINARY_RPM=1"
+        if [ $? -ne 0 ] ;
+        then
+            echo "cmake error"
+            return $?
+        fi
+    else
+        cmake .
     fi
+
     make package
     if [ $? -ne 0 ] ;
     then
@@ -258,38 +232,6 @@ build_rpm()
 
     return 0
 }
-
-
-build_rpm_renew()
-{
-    echo "renew rpm package start"
-
-    # build package directory
-    if [ ! -d "build_rpm" ] ; 
-    then
-        echo "error - directory build_rpm do not exist"
-        return 1;
-    fi
-    cd build_rpm
-
-    # build package
-   echo "${CPACK_BINARY}"
-    cmake .
-    if [ $? -ne 0 ] ;
-    then
-        echo "cmake error"
-        return $?
-    fi
-    make package
-    if [ $? -ne 0 ] ;
-    then
-        echo "make package error"
-        return $?
-    fi
-
-    return 0
-}
-
 
 
 build_rpm_clean()
@@ -299,24 +241,43 @@ build_rpm_clean()
 
 
 
+# -----------------------------------------------------------------------------
+# -----------------------------------------------------------------------------
+#
+# build test
+#
+# -----------------------------------------------------------------------------
+# -----------------------------------------------------------------------------
 build_tst()
 {    
     echo "build tst start"
 
     # build tst directory
-    rm -rf build_tst
-    mkdir build_tst
+    if [ ! -d "build_tst" ] ;
+    then
+        BUILD_FIRST=1
+        rm -rf build_tst
+        mkdir build_tst
+    else
+        BUILD_FIRST=0
+    fi
     cd build_tst
 
     # build tst
-    cmake ../tst \
-  	"${CMAKE_GENERATOR_LOCAL}" \
-	-DINSTALL_PREFIX_OpcUaStack="${HOME}/install"
-    if [ $? -ne 0 ] ;
+    if [ ${BUILD_FIRST} -eq 1 ] ;
     then
-        echo "cmake error"
-        return $?
+        cmake ../tst \
+  	     "${CMAKE_GENERATOR_LOCAL}" \
+	     -DINSTALL_PREFIX_OpcUaStack="${HOME}/install"
+        if [ $? -ne 0 ] ;
+        then
+            echo "cmake error"
+            return $?
+        fi
+    else
+        cmake .
     fi
+
     make 
      if [ $? -ne 0 ] ;
     then
@@ -330,6 +291,13 @@ build_tst()
 }
 
 
+# -----------------------------------------------------------------------------
+# -----------------------------------------------------------------------------
+#
+# cleanup
+#
+# -----------------------------------------------------------------------------
+# -----------------------------------------------------------------------------
 build_tst_clean()
 {
     rm -rf build_tst
@@ -351,59 +319,36 @@ clean()
 #
 # -----------------------------------------------------------------------------
 # -----------------------------------------------------------------------------
-if [ $# -eq 1 ] ; 
+if [ $# -ne 1 ] ; 
 then
-    RENEW=0
-elif [ $# -eq 2 ] ;
-then
-    if [ "$2" = "renew" ] ;
-    then
-        RENEW=1
-    else
-        usage
-        exit 1
-    fi
-else
     usage
     exit 1
 fi 
 
 if [ "$1" = "info" ] ;
 then
-    version_info
+    build_info
     exit $?
 elif [ "$1" = "clean" ] ;
 then 
     clean 
 elif [ "$1" = "local" ] ;
 then 
-    if [ ${RENEW} -eq 1 ] ;
-    then
-        build_local_renew
-    else
-        build_local
-    fi 
+    build_local
     exit $?
 elif [ "$1" = "deb" ] ;
 then 
-    if [ ${RENEW} -eq 1 ] ;
-    then
-        build_deb_renew
-    else
-        build_deb 
-    fi
+    build_deb 
     exit $?
 elif [ "$1" = "rpm" ] ;
 then 
-    if [ ${RENEW} -eq 1 ] ;
-    then
-        build_rpm_renew
-    else 
-        build_rpm
-    fi 
+    build_rpm
+    exit $?
+elif [ "$1" = "tst" ] ;
+then 
+    build_tst
     exit $?
 else
     usage
+    exit 1
 fi
-
-exit 0
